Parsowanie wiecej niz 1 nawiasu zamykajacego, porzadki

+++ src/ParserPQL/MatcherPQL.h
@@ -63,7 +63,7 @@ public:
 		int pos2 = elem.find("\"",pos1+1);
 		int pos3 = elem.find("\"",pos2+1);
 
-		if(pos1 < pos2 && pos1 + 1 != pos2 && pos3 > elem.length())
+		if(pos1 < pos2 && pos1 + 1 != pos2 && pos3 > (int) elem.length())
 			return true;
 		else
 			return false;
@@ -71,7 +71,7 @@ public:
 
 	bool isStar(string elem, int position)
 	{
-		if(elem.find("*",position) < elem.length() && elem.find("*",position) == position)
+		if(elem.find("*",position) <  elem.length() && (int) elem.find("*",position) == position)
 			return true;
 		else
 			return false;
@@ -82,7 +82,7 @@ public:
 		int pos1 = elem.find(",");
 		int pos2 = elem.find(",",pos1+1);
 
-		if(pos1 < elem.length() && pos2 > elem.length())
+		if(pos1 < (int) elem.length() && pos2 > (int) elem.length())
 			return true;
 		else
 			return false;
@@ -93,7 +93,7 @@ public:
 		int pos1 = elem.find("(");
 		int pos2 = elem.find(")");
 
-		if(pos1 == 0 && pos2 == elem.length() - 1)
+		if(pos1 == 0 && pos2 == (int) elem.length() - 1)
 			return true;
 		else
 			return false;
@@ -114,7 +114,7 @@ public:
 
 	bool isNumber(string elem)
 	{
-		for(int i = 0 ; i < elem.length() ; i ++)
+		for(unsigned i = 0 ; i < elem.length() ; i ++)
 		{
 			if(!isdigit(elem[i]))
 				return false;
